fix(worktree): widen git-common watch on event-batch overflow
The watcher child drops the whole batch and reports {op: 'overflow'}
when a burst exceeds 5,000 events (e.g. a fleet-wide bulk op), but
startGitCommonNarrowWatch never consumed onOverflow — a silent loss
mode with no other signal, unlike interruption/error/root-deletion
which were already covered. Wire it through to a dedicated
handleWatchOverflow path that widens unconditionally (structural +
status + head-identity) and bypasses the watcher-error refresh
cooldown, since overflow is definite proof of loss and a single bulk
op can legitimately overflow more than once before settling.

Split worktree-base-directory-poller.ts (the marker-based fallback
poller) and worktree-base-directory-watcher.ts (the watch-event
handlers) into dedicated files to stay under the max-lines gate.

Refs #17828
